requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.24.0
matplotlib>=3.7.0
tqdm>=4.65.0
opencv-python>=4.8.0
streamlit>=1.31.0
streamlit-drawable-canvas>=0.9.3
